// types/github-content.d.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface GitHubFileType {
  name: string;
  path: string;
  sha: string;
  size: number;
  url: string;
  html_url: string;
  git_url: string;
  download_url: string;
  type: string;
  _links: Link;
  commit: Commit;
}

export interface Link {
  self: string;
  git: string;
  html: string;
}

export interface Commit {
  author: Author;
  committer: Author;
  message: string;
  tree: Tree;
  url: string;
  comment_count: number;
  verification: Verification;
}

export interface Author {
  name: string;
  email: string;
  date: string;
}

export interface Tree {
  sha: string;
  url: string;
}

export interface Verification {
  verified: boolean;
  reason: string;
  signature: null;
  payload: null;
}

interface FileExistsData {
  sha?: string;
  contentSha?: string;
  metadataSha?: string;
  treeSha?: string;
  parentSha?: string;
  commitSha?: string;
}